requirements.txt
# NextStep Backend Dependencies - Updated: 2025-01-15
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# LangChain ecosystem - let pip resolve compatible versions
langchain>=0.1.20,<0.3.0
langchain-openai>=0.1.8,<0.2.0
langchain_chroma>=0.1.2,<0.2.0
langchain-text-splitters>=0.0.2,<0.3.0

# Vector store and embeddings - compatible ranges
chromadb>=0.4.24,<0.6.0
openai>=1.35.0,<2.0.0

# Document processing
pypdf==3.17.4
beautifulsoup4==4.12.2

# Core dependencies
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<3.0.0
aiofiles>=23.0.0,<24.0.0

# Additional dependencies for Railway deployment
pydantic>=2.4.0,<3.0.0
typing-extensions>=4.8.0

# Google APIs for two-way client communication
google-api-python-client>=2.100.0,<3.0.0
google-auth-httplib2>=0.2.0,<1.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth>=2.20.0,<3.0.0

# Email, SMS, and Voice services
emails>=0.6.0,<1.0.0
sendgrid>=6.10.0,<7.0.0
twilio>=8.10.0,<9.0.0